requirements.txt
Django==4.2
djangorestframework==3.14.0
poetry==1.7.0
djoser==2.2.2
python-dotenv==1.0.0
Pillow==10.1.0
psycopg2-binary==2.9.9